+++ src/main/groovy/io/xh/hoist/role/RoleMemberChanges.groovy
@@ -1,0 +1,2 @@
+package io.xh.hoist.role
+